workaround java 1.5 bug where process.waitFor() only returns values 0..255
git-svn-id: https://svn.mpi-sws.org/DS/freepastry/trunk@3220 c21dacaf-8602-0410-aacd-8bc41df2cbe7

+++ pastry/src/rice/proxy/Proxy.java
@@ -70,7 +70,9 @@ public class Proxy {
       
       Printer error = new Printer(process.getErrorStream(), "[Error Stream ]: ");
       
-      int exit = process.waitFor();    
+      // have to do the byte cast below because java 5 only returns low 8 bits
+      // probably we should change all our status codes to be 0..255 some day
+      int exit = (byte)process.waitFor();    
       lm.die();
       
       // re-initialize parameters for debugging purposes
